clean code EventEmitter
split all method
