+++ methods/arrayClone.js
@@ -1,0 +1,6 @@
+module.exports = function arrayClone(arr, n) {
+  var copy = new Array(n);
+  for (var i = 0; i < n; ++i)
+    copy[i] = arr[i];
+  return copy;
+}
