+++ methods/bind.js
@@ -1,0 +1,6 @@
+module.exports = Function.prototype.bind || function functionBindPolyfill(context) {
+  var fn = this;
+  return function () {
+    return fn.apply(context, arguments);
+  };
+};
