+++ methods/emitMany.js
@@ -1,0 +1,12 @@
+var arrayClone = require('./arrayClone');
+
+module.exports = function emitMany(handler, isFn, self, args) {
+  if (isFn)
+    handler.apply(self, args);
+  else {
+    var len = handler.length;
+    var listeners = arrayClone(handler, len);
+    for (var i = 0; i < len; ++i)
+      listeners[i].apply(self, args);
+  }
+}
