+++ methods/emitNone.js
@@ -1,0 +1,12 @@
+var arrayClone = require('./arrayClone');
+
+module.exports = function emitNone(handler, isFn, self) {
+  if (isFn)
+    handler.call(self);
+  else {
+    var len = handler.length;
+    var listeners = arrayClone(handler, len);
+    for (var i = 0; i < len; ++i)
+      listeners[i].call(self);
+  }
+}
