+++ methods/emitOne.js
@@ -1,0 +1,12 @@
+var arrayClone = require('./arrayClone');
+
+module.exports = function emitOne(handler, isFn, self, arg1) {
+  if (isFn)
+    handler.call(self, arg1);
+  else {
+    var len = handler.length;
+    var listeners = arrayClone(handler, len);
+    for (var i = 0; i < len; ++i)
+      listeners[i].call(self, arg1);
+  }
+}
