+++ methods/eventNames.js
@@ -1,0 +1,3 @@
+module.exports = function eventNames() {
+  return this._eventsCount > 0 ? Reflect.ownKeys(this._events) : [];
+};
