+++ methods/objectCreate.js
@@ -1,0 +1,5 @@
+module.exports = Object.create || function objectCreatePolyfill(proto) {
+  var F = function () { };
+  F.prototype = proto;
+  return new F;
+};
