+++ methods/objectKeys.js
@@ -1,0 +1,7 @@
+module.exports = Object.keys || function objectKeysPolyfill(obj) {
+  var keys = [];
+  for (var k in obj) if (Object.prototype.hasOwnProperty.call(obj, k)) {
+    keys.push(k);
+  }
+  return k;
+};
