+++ methods/onceWrap.js
@@ -1,0 +1,10 @@
+var bind = require('./bind');
+var onceWrapper = require('./onceWrapper');
+
+module.exports = function onceWrap(target, type, listener) {
+  var state = { fired: false, wrapFn: undefined, target: target, type: type, listener: listener };
+  var wrapped = bind.call(onceWrapper, state);
+  wrapped.listener = listener;
+  state.wrapFn = wrapped;
+  return wrapped;
+}
