+++ methods/setMaxListeners.js
@@ -1,0 +1,6 @@
+module.exports = function setMaxListeners(n) {
+  if (typeof n !== 'number' || n < 0 || isNaN(n))
+    throw new TypeError('"n" argument must be a positive number');
+  this._maxListeners = n;
+  return this;
+};
