+++ methods/spliceOne.js
@@ -1,0 +1,5 @@
+module.exports = function spliceOne(list, index) {
+  for (var i = index, k = i + 1, n = list.length; k < n; i += 1, k += 1)
+    list[i] = list[k];
+  list.pop();
+}
