+++ methods/unwrapListeners.js
@@ -1,0 +1,7 @@
+module.exports = function unwrapListeners(arr) {
+  var ret = new Array(arr.length);
+  for (var i = 0; i < ret.length; ++i) {
+    ret[i] = arr[i].listener || arr[i];
+  }
+  return ret;
+}
